[FEATURE] Pre rendering scripts and GitHub workflow (#4)
A GitHub action will now run pre-rendering scripts and push changes to `main` whenever changes are pushed to the `contrib` branch.

These commands can also be run with `make`. `docs` should still be the main argument for rendering documentation locally.

* `cleanup-docs`: Restores the Documentation folder from Documentation-BACKUP-temp
* `docs`: Prepare, generate, and clean up docs.
* `expand-tags`: Expands tags into Markdown
* `generate-tos`: Generates table of contents in all Index.md files
* `help`: Displays this list of targets with descriptions
* `index-files`: Indexes all files and links in files.json
* `expand-links`: Expands shorthand links
* `render-docs`: Render docs base on Documentation directory
* `stage-docs`: Make a backup of the Documentation directory in Documentation-BACKUP-temp
* `test-docs`: Test the documentation rendering

Also implements a search-replace for shorthand links to only a file's file name. Since each step-by-step guide should have a unique title, it follows that a file name based on the title should also be unique. This allows us to link to a file by only its name, and we don't have to think about the folder structure.